+++ client/src/reducers/user.js
@@ -1,0 +1,72 @@
+import { USER_SUBMIT_EMAIL, RECEIVE_USER_INFO } from '../actions/user_signup';
+import { LOGIN_SUBMIT, LOGIN_ERROR, LOGIN_SUCCESSFUL } from '../actions/login';
+import { CALLING_NOW, CALL_ERROR, CALL_SENT } from '../actions/call.js';
+
+const initialState = {
+  id: null,
+  firstName: '',
+  lastName: '',
+  phone: '',
+  email: '',
+  isLoggingIn: false,
+  isCalling: false,
+  error: null
+};
+
+export default function user(state=initialState, action) {
+  switch (action.type) {
+    case USER_SUBMIT_EMAIL:
+      return {
+        ...state,
+        email: action.email
+      }
+    case RECEIVE_USER_INFO:
+      return {
+        ...state,
+        id: action.id,
+        firstName: action.firstName,
+        lastName: action.lastName,
+        phone: action.phone,
+      }
+    case LOGIN_SUBMIT:
+      return {
+        ...state,
+        isLoggingIn: true
+      }
+    case LOGIN_ERROR:
+      return {
+        ...state,
+        isLoggingIn: false,
+        error: action.error
+      }
+    case LOGIN_SUCCESSFUL:
+      return {
+        ...state,
+        id: action.id,
+        firstName: action.firstName,
+        lastName: action.lastName,
+        phone: action.phone,
+        email: action.email,
+        isLoggingIn: false,
+        error: null
+      }
+    case CALLING_NOW:
+      return {
+        ...state,
+        isCalling: true
+      }
+    case CALL_ERROR:
+      return {
+        ...state,
+        isCalling: false,
+        error: action.error
+      }
+    case CALL_SENT:
+      return {
+        ...state,
+        isCalling: false
+      }
+    default:
+      return state;
+  }
+}
